Track "good"/approved releases of libraries
Use an on-wiki page (for history/diffs/auth/etc.) to identify releases
of packages as good for libraryupgrader to go ahead with updating
instead of requiring manual intervention from a human.

Whatever uses this should also regularly check it for when a release is
pulled/de-approved.

Theoretically in the future we could use gpg signed tags for at least
composer, but that infrastructure seems to be crumbling for now. :(

Change-Id: I649e6ae26aaaff51259a62a4a7fde75bc3a6b523