Adjust eslint config to enable caching and reporting extra disables
The parser in grunt.py is good enough to parse every Gruntfile.js in
Gerrit except Popups because it has a comment (meh). For the most part
it roundtrips well, though sometimes it'll split a list into multiple
lines when it was on one previously (or the other way around).

Good enough.

Bug: T225325
Change-Id: Ic88ba15ff5c3b7ba0f25b02fa117c259b6fa2327