Introduce diff-libraries
For a Gerrit patch, it produces a real diff of any changes made to the
various libraries that are depended upon.

The flow is:
* Download patch
* See if composer.json/package.json/package-lock.json were modified
* If so, run composer install/npm ci, copy to another directory
* Clean up, checkout HEAD~1, run install stuff again
* Generate a diff and save it

It also takes a few steps to reduce unnecessary variance, like setting
composer's autoloader-suffix to a static value instead of relying on the
default randomness. The same folder is used for installing because npm
hardcodes the directory path inside of node_modules for every single
package that's installed.

From a security point of view there are quite a few holes in this
implementation. We execute npm/composer in the same permissions context
as the web app and database, though it should be straightforward to fix
later. The biggest issue is that we have no protection against
Volkswagen-style maliciousness since we're diffing only the files on
disk, not what is being executed. Also it tries to do everything in
parallel (yay, a language with good concurrency!) so it should be
straightforward to DoS.

This will be deployed as a docker image. For now it's going to be an
emphemeral service, in that the database will get wiped on every
deploy/restart. It'll go on a separate domain and instance from libup,
just because I'm not 100% confident with the isolation.

Change-Id: Ia6873068eecaba56074cbb9f25e943d981cdf1ff