Keep a mirror of git clones so we're not constantly doing fresh ones
For every repository access, we do a shallow git clone from the Gerrit
replica. We can be more efficient (especially with branch support coming
soon) by keeping a full, bare clone in /srv/git and cloning from there
as needed.

Note that this does not apply to the labs/libraryupgrader/config repo,
which continues to have its own special handling.

Change-Id: I205577a6c55a6ecfae4f76dbef6b6f3f699b4462