Use the new plan/config v2 to decide which updates to implement
The container will make a POST request to the webserver, which responds
with a JSON list of updates to do. Because the database has the latest
dependencies right before the container runs, it should all be in sync.

One change is that the weight is now calculated in the container, mainly
so we can see what an in progress but not pushed patch would look like.

Change-Id: I407962910e8610cc934d6bce7de12ade64f94bc0