Update eslint with --depth 10
Some packages will have a nested eslint (via grunt-eslint) on an older
version, this tells npm to update all the way to a depth of 10. In theory
we'd use a very high number, like 9999 (as suggested in the npm docs), but
that triggers "npm ERR! Maximum call stack size exceeded".

Change-Id: I666f857b75363e2806251b1c201c92305bd58a38