Add Phabricator release monitoring bot
Currently using release-monitoring.org since their API allows for
unauthenticated access.

Versions and previously filed tasks are stored in the "monitoring"
database table. If the task is still open, a new comment will be added
in the case of a new version. If it's closed or a task was never filed,
a new task will be created.

In the future we should probably also add a web endpoint dumping the
status of all the checks.

Bug: T205365
Change-Id: I62e1c80825d876c7aa19dccbfb5f6d69fe94721b