Support "main" branch
In the LibUp web UI, the "master" branch will look as if it is "main",
but all Git operations will still use the correct branch of "master" (or
"main").

The database now stores the LibUp branch ("master" is normalized to
"main") and a separate Git branch. I audited every use of
`repo.branch` to make sure it uses the correct one.

The runner only knows about the git branch, so we need to handle
normalization there (limited to is_main()) and in the one HTTP API
it hits, plan.json.

There's some undefined behavior if a repository has both "master" and
"main" branches...I hope that doesn't happen.

Bug: T273237
Change-Id: Iaec9992f3205347f4df7a9d9dcab469cff316a5c