Switch runner to bullseye, Python 3.9 and npm 7
...and fix a bunch of overdue things.

The main thing here is that the npm audit format has entirely changed
and is very stripped down of most information. All advisories now link
to GitHub Security Advisories, so we lose CVEs and plenty of other
useful information. Sigh, this is what we get for using a package
manager and registry controlled by a proprietary company.

With bullseye we can now use the packaged npm and composer, which makes
it trivial for us to guard against CVE-2020-5252 style attacks since
those files are owned by root and shouldn't be able to be tampered with
(this was also really only an issue for npm to begin with).

There are probably other Python 3.9 improvements we can make, but that's
left for a future commit.

Give up on using hashes in requirements.txt because for some reason it
doesn't work and hopefully we can drop all this indirection by using
a PEP 518 build system soon. Or something.

Bug: T273810
Bug: T282307
Bug: T228173
Bug: T228157
Bug: T249525
Change-Id: I8d03cb204b4a1a1b499b08a636d68e37ad4b4be6